chore: cleanup and restructure
